refactor: combine types
