+++ src/chains/defaults/messages/getLiquidationMessage.ts
@@ -2,11 +2,13 @@ import { toUtf8 } from "@cosmjs/encoding";
 import { EncodeObject } from "@cosmjs/proto-signing";
 import { MsgExecuteContract } from "cosmjs-types/cosmwasm/wasm/v1/tx";
 
+import { LiquidationMessage } from "../../../core/types/messages/liquidationmessages";
+
 /**
  *
  */
 export function getliqudationMessage(sender: string, overseerAddress: string, borrowerAddress: string) {
-	const message = {
+	const message: LiquidationMessage = {
 		liquidate_collateral: {
 			borrower: borrowerAddress,
 		},
