--- src/core/types/messages/pricefeedmessage.ts
@@ -1,5 +1,0 @@
-export interface PriceFeedMessage {
-	feed_price: {
-		prices: Array<[string, string]>;
-	};
-}
